Moving parser functions to their own file
